Se inicia crud de cursos

+++ app/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render,redirect, get_object_or_404
-from .models import Alumno
+from .models import Alumno, Cursos
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib.auth import authenticate, login as auth_login
@@ -23,7 +23,6 @@ def login(request):
 
 def alumnos(request):
     if request.method == 'GET':
-        print("alumno: ",Alumno.objects.all())
         return render(request, 'app/alumnos.html',{'alumnos': Alumno.objects.all()})
     else:
         return render(request, 'app/alumnos.html',{'alumnos': Alumno.objects.all()})
@@ -53,6 +52,7 @@ def alumno_agregar(request):
 def alumno_editar(request, id):
     #alumno = Alumno.objects.get(id=id)
     alumno = get_object_or_404(Alumno, id=id)
+    
     if request.method == 'POST':
         alumno.nombre=request.POST["nombre"]
         alumno.apellidos=request.POST["apellidos"]
@@ -67,7 +67,58 @@ def alumno_editar(request, id):
         return render(request, 'app/alumno_editar.html', {'alumno': alumno})
     
 def alumno_eliminar(request, id):
-    alumno = Alumno.objects.get(id=id)
-    alumno.delete()
-    return redirect('alumnos')
+    if request.method == 'POST':
+        alumno = Alumno.objects.get(id=id)
+        alumno.delete()
+        return redirect('alumnos')
+    # else:
+    #     return render(request, 'app/alumno_eliminar.html', {'alumno': alumno})
 
+# -------------- Cursos --------------
+def cursos(request):
+    if request.method == 'GET':
+        return render(request, 'app/cursos.html',{'cursos': Cursos.objects.all()})
+    else:
+        return render(request, 'app/cursos.html',{'cursos': Cursos.objects.all()})
+    
+
+def curso_agregar(request):
+    if request.method == 'POST':
+        curso=Cursos(
+            nombre=request.POST["nombre"],
+            apellidos=request.POST["apellidos"],
+            edad=request.POST["edad"],
+            correo=request.POST["correo"],
+            numDoc=request.POST["numDoc"],
+            usuarioCreacion=request.user,
+            estado=request.POST["estado"]
+            )
+        # grabar producto en la BD
+        curso.save()
+        return redirect('cursos')
+    else:
+        return render(request, 'app/curso_agregar.html')
+    
+def curso_editar(request, id):
+    curso = get_object_or_404(Cursos, id=id)
+    
+    if request.method == 'POST':
+        curso.nombre=request.POST["nombre"]
+        curso.descripcion=request.POST["descripcion"]
+        curso.fechaInicio=request.POST["fechaInicio"]
+        curso.fechaFin=request.POST["fechaFin"]
+        curso.usuarioCreacion=request.user    
+        curso.estado=request.POST["estado"]
+        curso.save()
+        return redirect('cursos')
+    else:
+        fi = curso.fechaInicio
+        ff = curso.fechaFin
+        curso.fechaInicio = '{:%Y-%m-%d}'.format(fi)
+        curso.fechaFin = '{:%Y-%m-%d}'.format(ff)
+        return render(request, 'app/curso_editar.html', {'curso': curso})
+    
+def curso_eliminar(request, id):
+    curso = Cursos.objects.get(id=id)
+    curso.delete()
+    return redirect('cursos')
